Adds JSON schema validation to config files
Adds a JSON schema file for validating portmux configuration files.

This change introduces a schema to improve config file maintainability and
provide better editor support through autocompletion and validation.

The schema is included in the CLI package and referenced in the default
configuration file, and the init command. It also updates the core schema
to allow for the `$schema` key.